chore(config): move JWT configuration to bottom of application.yml

+++ src/main/resources/db/migration/initial/002_create_users.sql
@@ -10,7 +10,7 @@ CREATE TABLE users (
     email          VARCHAR(255) UNIQUE NOT NULL,
     password       VARCHAR(255) NOT NULL,
     enabled        BOOLEAN NOT NULL DEFAULT TRUE,
-    authorities_id INTEGER,
+    authorities_id BIGINT,
 
     CONSTRAINT fk_user_authorities
         FOREIGN KEY (authorities_id)
